SCHEDULER: Move scheduler code to a separate directory
