+++ src/main/io/serial_msp.c
@@ -24,7 +24,8 @@
 #include "build_config.h"
 #include "debug.h"
 #include "platform.h"
-#include "scheduler.h"
+
+#include "scheduler/scheduler.h"
 
 #include "common/axis.h"
 #include "common/color.h"
