+++ src/main/sensors/barometer.c
@@ -20,7 +20,8 @@
 #include <math.h>
 
 #include "platform.h"
-#include "scheduler.h"
+
+#include "scheduler/scheduler.h"
 
 #include "common/maths.h"
